requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP & Data
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Image Generation
Pillow==10.2.0
cairosvg==2.7.1

# Database
redis==5.0.1
asyncpg==0.29.0

# AI & NLP
openai==1.10.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Farcaster
farcaster-py==0.1.0

# Additional
pytz==2023.3
